updated build script for windows

+++ build_exe.ps1
@@ -1,8 +1,12 @@
-$DistDirectory = "./dist"
-if (Test-Path $DistDirectory) {
-  Remove-Item $DistDirectory -Recurse
-}
+# Assume we are already in a conda environment
 
-conda activate oct_labeler
+# if dist exists, remove it
+$DistDirectory = "./dist"
+if (Test-Path $DistDirectory) { Remove-Item $DistDirectory -Recurse }
+
+# Install oct_labeler
 pip install -e .
+if (!$?) { Exit $LASTEXITCODE }
+
+# build binary
 pyinstaller main.spec
